Reshape SDK public world surface
Introduce dedicated block and volumetric SDK type crates, decompose freven_world_api into focused modules, and update guest-facing crates and workspace metadata to match the new public surface boundaries.
