+++ Cargo.toml
@@ -9,6 +9,8 @@ members = [
     "crates/freven_world_guest",
     "crates/freven_world_guest_sdk",
     "crates/freven_world_sdk_types",
+    "crates/freven_volumetric_sdk_types",
+    "crates/freven_block_sdk_types",
 ]
 
 [workspace.package]
@@ -27,6 +29,8 @@ freven_guest_sdk = { path = "crates/freven_guest_sdk", version = "0.1.0" }
 freven_sdk_types = { path = "crates/freven_sdk_types", version = "0.1.0" }
 freven_world_guest = { path = "crates/freven_world_guest", version = "0.1.0" }
 freven_world_sdk_types = { path = "crates/freven_world_sdk_types", version = "0.1.0" }
+freven_volumetric_sdk_types = { path = "crates/freven_volumetric_sdk_types", version = "0.1.0" }
+freven_block_sdk_types = { path = "crates/freven_block_sdk_types", version = "0.1.0" }
 postcard = { version = "1.1.3", default-features = false, features = ["alloc"] }
 serde = { version = "1.0.228", features = ["derive"] }
 toml = "1.0.6"
