+++ crates/freven_world_api/Cargo.toml
@@ -16,6 +16,8 @@ freven_mod_api.workspace = true
 freven_sdk_types.workspace = true
 freven_world_guest.workspace = true
 freven_world_sdk_types.workspace = true
+freven_volumetric_sdk_types.workspace = true
+freven_block_sdk_types.workspace = true
 serde.workspace = true
 toml.workspace = true
 thiserror.workspace = true
