+++ crates/freven_world_guest/src/lib.rs
@@ -12,7 +12,7 @@ use freven_guest::{
     CapabilityDeclaration, ChannelDeclaration, ComponentDeclaration, LifecycleHooks, LogPayload,
     MessageDeclaration, MessageHooks, RuntimeSessionInfo,
 };
-use freven_world_sdk_types::blocks::{BlockDescriptor, BlockRuntimeId};
+use freven_block_sdk_types::{BlockDescriptor, BlockRuntimeId};
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
