+++ crates/freven_world_guest_sdk/Cargo.toml
@@ -10,6 +10,7 @@ freven_guest.workspace = true
 freven_sdk_types.workspace = true
 freven_world_guest.workspace = true
 freven_world_sdk_types.workspace = true
+freven_block_sdk_types.workspace = true
 postcard.workspace = true
 serde.workspace = true
 toml.workspace = true
